[up:objective-functions-losses] refactor: leave only the oo api (enhanced)
