--- nujo/objective/__init__.py
@@ -1,2 +1,0 @@
-from nujo.objective.functional import *
-from nujo.objective.object_oriented import *
